Extract spawn zone logic to partial form
